Windows compatibility

+++ libs/client-sdk/src/data_sources/base_64.cpp
@@ -59,7 +59,7 @@ std::string Base64UrlEncode(std::string const& input) {
     while (bit_index < bit_count) {
         // Get either 6 bits, or the remaining number of bits.
         auto bits = GetBits(bit_index,
-                            std::min(kIndexBits, bit_count - bit_index), input);
+                            std::min(kIndexBits, static_cast<unsigned long>(bit_count - bit_index)), input);
         out.push_back(static_cast<char>(kBase64Table.at(bits.to_ulong())));
         bit_index += kIndexBits;
     }
